enhanced scripts: - added record button - added prefabs for pedestrian - changed rendering for obstacles

+++ Assets/Scripts/Pedestrian/PedestrianPosition.cs
@@ -11,11 +11,11 @@ using System;
 
 public class PedestrianPosition {
 	private int flooID;
-	private decimal time;
+	private float time;
 	private float x;
 	private float y;
     private float z;
-	public PedestrianPosition (int id, decimal time, float x, float y, float z) {
+	public PedestrianPosition (int id, float time, float x, float y, float z) {
 		this.flooID = id;
 		this.time = time;
 		this.x = x;
@@ -24,7 +24,7 @@ public class PedestrianPosition {
 	}
 
 	public int getFloorID() {return this.flooID;}
-	public decimal getTime() {return this.time;}
+	public float getTime() {return this.time;}
 	public float getX() {return this.x;}
 	public float getY() {return this.y;}
     public float getZ() {return this.z;}
